requirements.txt
autopep8==1.5.2
bcrypt==3.1.7
cffi==1.14.0
click==7.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.1.2
pycodestyle==2.5.0
pycparser==2.20
six==1.14.0
SQLAlchemy==1.3.16
Werkzeug==1.0.1
WTForms==2.2.1
